requirements.txt
python-dotenv
pyyaml
pydantic
pydantic-settings
pandas
openai
qdrant-client
openpyxl>=3.1.0
pymupdf4llm==0.0.27
streamlit
pytest
ruff
mypy
types-PyYAML

